Move SharedInverter to library
