+++ .gitignore
@@ -3,6 +3,9 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
+# Git generated
+*.orig
+
 # User-specific files
 *.suo
 *.user
@@ -315,25 +318,14 @@ ssc/examples/solar_pilot_example.lk
 set_env_sj.bat
 
 build_osx/dltest\.dSYM/Contents/
-
 build_osx/SDKtool\.app/Contents/
-
 build_osx/ssc\.dylib
-
 build_osx/TCSConsole\.app/Contents/
-
 build_ios/libssc\.zip
-
 build_android/libssc\.zip
-
 android_make_output\.txt
-
 ios_make_output\.txt
-
 build_osx/Test
-
 build_linux/Test
-
 build_linux/config.status
-
 linux_test\.txt
