+++ build_vs2017/shared_vs2017.vcxproj
@@ -43,6 +43,7 @@
     <ClInclude Include="..\shared\lib_pvwatts.h" />
     <ClInclude Include="..\shared\lib_pv_shade_loss_mpp.h" />
     <ClInclude Include="..\shared\lib_sandia.h" />
+    <ClInclude Include="..\shared\lib_shared_inverter.h" />
     <ClInclude Include="..\shared\lib_snowmodel.h" />
     <ClInclude Include="..\shared\lib_util.h" />
     <ClInclude Include="..\shared\lib_utility_rate.h" />
@@ -72,6 +73,7 @@
     <ClCompile Include="..\shared\lib_pvwatts.cpp" />
     <ClCompile Include="..\shared\lib_pv_shade_loss_mpp.cpp" />
     <ClCompile Include="..\shared\lib_sandia.cpp" />
+    <ClCompile Include="..\shared\lib_shared_inverter.cpp" />
     <ClCompile Include="..\shared\lib_snowmodel.cpp" />
     <ClCompile Include="..\shared\lib_util.cpp" />
     <ClCompile Include="..\shared\lib_utility_rate.cpp" />
